playwright ui tests

+++ db_models/movie_db_model.py
@@ -10,7 +10,7 @@ class MovieDBModel(Base):
 
     id = Column(Integer, primary_key=True)
     name = Column(String)
-    price = Column(Float)
+    price = Column(Integer)
     description = Column(String)
     image_url = Column(String)
     location = Column(String)
